requirements.txt
streamlit
torch
whisper
deep-translator
groq
gTTS
pygame
python-dotenv
requests
